refactor(worker): extract pure row transforms from repository.js
P4 U9: move 29 genuinely pure row-transform functions and 9 associated
constants from repository.js (9,322 lines) into a new focused module
worker/src/row-transforms.js (394 lines). repository.js drops to 9,077
lines (-245 net).

Extracted functions (row in -> record out, no db, no service, no async):
  subjectStateRowToRecord, gameStateRowToRecord, publicMonsterCodexEntry,
  publicMonsterCodexState, publicGameStateRowToRecord,
  publicMonsterCodexStateFromSpellingProgress, publicMonsterCodexHasMastery,
  secureSpellingProgress, spellingProgressFromSubjectRow,
  practiceSessionRowToRecord, publicPracticeSessionRowToRecord,
  publicPracticeLabel, publicSummaryCards, publicMistakeSummary,
  publicPracticeSessionSummary, publicPunctuationPracticeSessionSummary,
  eventRowToRecord, publicEventRowToRecord, contentRowToBundle,
  safeSpellingPrompt, safeSpellingCurrentCard, safeSpellingSessionProgress,
  publicSpellingStats, publicSpellingAnalytics,
  safePublicEventText, safePublicEventNumber, safePublicEventType,
  safePublicEventEnum

Extracted constants:
  PUBLIC_SPELLING_YEAR_LABELS, PUBLIC_PRACTICE_CARD_LABELS,
  PUBLIC_EVENT_TYPES, PUBLIC_MONSTER_CODEX_SYSTEM_ID, PUBLIC_MONSTER_IDS,
  PUBLIC_DIRECT_SPELLING_MONSTER_IDS, PUBLIC_MONSTER_BRANCHES,
  SPELLING_SECURE_STAGE, PUBLIC_EVENT_TEXT_ENUMS

NOT extracted (impure — plan feasibility constraints):
  redactPunctuationUiForClient (instantiates createPunctuationService)
  publicSubjectStateRowToRecord (async, builds audio cues)
  mergePublicSpellingCodexState (takes db parameter)

Barrel re-exports at the bottom of repository.js follow the P3 pattern
so all existing consumers are unaffected.

Zero behaviour change. 325+ tests verified across bootstrap, commands,
subject runtime, admin, hubs, projections, capacity, and query budget
suites.
